backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "numpy>=2.2.4",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.24",
    "torch>=2.6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.9.9",
    "taskipy>=1.14.1",
]

[tool.ruff]
include = ["pyproject.toml", "app/**/*.py", "tests/**/*.py", "utils/**/*.py"]

[tool.taskipy.tasks]
start = "uv run uvicorn app.main:app"
dev = "uv run uvicorn app.main:app --reload"
lint = "uv run ruff check"
test = "uv run pytest -q"
test_debug = "uv run pytest -s"